Implement asynchronous haul cycles computation

+++ processes/src/main/java/com/github/vkremianskii/pits/processes/model/EquipmentPayloadRecord.java
@@ -28,7 +28,7 @@ public class EquipmentPayloadRecord {
         return equipmentId;
     }
 
-    public double payload() {
+    public int payload() {
         return payload;
     }
